Move ChannelsListWithData to separate file
