+++ client/src/App.js
@@ -1,10 +1,9 @@
 import React, { Component } from 'react';
 import logo from './apollo-logo.svg';
 import './App.css';
+import ChannelsListWithData from './components/ChannelsListWithData';
 
 import {
-  graphql,
-  gql,
   ApolloClient,
   ApolloProvider,
 } from 'react-apollo';
@@ -22,32 +21,6 @@ const mockNetworkInterface = mockNetworkInterfaceWithSchema({ schema });
 const client = new ApolloClient({
   networkInterface: mockNetworkInterface,
 });
-
-const ChannelsList = ({ data: {loading, error, channels }}) => {
-  if (loading) {
-    return <p>Loading ...</p>;
-  }
-  if (error) {
-    return <p>{error.message}</p>;
-  }
-
-  return (
-    <div className="channelsList">
-      { channels.map( ch => <div key={ch.id} className="channel">{ch.name}</div> ) }
-    </div>
-  );
-};
-
-const channelsListQuery = gql`
-  query ChannelsListQuery {
-    channels {
-      id
-      name
-    }
-  }
-`;
-
-const ChannelsListWithData = graphql(channelsListQuery)(ChannelsList);
 
 class App extends Component {
   render() {
